+++ src/api.rs
@@ -9,7 +9,7 @@ use std::{
 use axum::{
     extract::{Path, Query, State},
     http::StatusCode,
-    routing::{get, post},
+    routing::{delete, get, post},
     Json, Router,
 };
 use log::{debug, error};
@@ -19,7 +19,7 @@ use crate::{
     crypto::PublicKey,
     endpoint::Endpoint,
     message::{MessageId, MessageInfo, MessageStack},
-    peer_manager::{PeerExists, PeerManager, PeerStats},
+    peer_manager::{PeerExists, PeerManager, PeerNotFound, PeerStats},
 };
 
 /// Default amount of time to try and send a message if it is not explicitly specified.
@@ -105,6 +105,7 @@ impl Http {
         let admin_routes = Router::new()
             .route("/admin", get(get_info))
             .route("/admin/peers", get(get_peers).post(add_peer))
+            .route("/admin/peers/:endpoint", delete(delete_peer))
             .route("/admin/routes/selected", get(get_selected_routes))
             .route("/admin/routes/fallback", get(get_fallback_routes))
             .with_state(server_state.clone());
@@ -352,6 +353,26 @@ async fn add_peer(
         Err(PeerExists) => Err((
             StatusCode::CONFLICT,
             "A peer identified by that endpoint already exists".to_string(),
+        )),
+    }
+}
+
+/// remove an existing peer from the system
+async fn delete_peer(
+    State(state): State<HttpServerState>,
+    Path(endpoint): Path<String>,
+) -> Result<StatusCode, (StatusCode, String)> {
+    debug!("Attempting to remove peer {} to  the system", endpoint);
+    let endpoint = match Endpoint::from_str(&endpoint) {
+        Ok(endpoint) => endpoint,
+        Err(e) => return Err((StatusCode::BAD_REQUEST, e.to_string())),
+    };
+
+    match state.peer_manager.delete_peer(&endpoint) {
+        Ok(()) => Ok(StatusCode::NO_CONTENT),
+        Err(PeerNotFound) => Err((
+            StatusCode::NOT_FOUND,
+            "A peer identified by that endpoint does not exist".to_string(),
         )),
     }
 }